requirements.txt
discord.py>=2.3.2
aiosqlite>=0.19.0
python-dateutil>=2.8.2
typing-extensions>=4.7.1
